requirements.txt
python-dotenv
flask
pymongo
torch
transformers